+++ sha256.h
@@ -11,7 +11,7 @@
 #define _SH256_H_
 
 /*************************** HEADER FILES ***************************/
-#include <stddef.h>
+#include <stdint.h>
 
 #include "cryptoalg_data_type.h"
 
@@ -30,10 +30,10 @@
  *      定義SHA256摘要算法上下文機構體
  */
 typedef struct {
-    BYTE data[64];
-    WORD datalen;
-    unsigned long long bitlen;
-    WORD state[8];
+    uint8_t     data[64];
+    uint32_t    datalen;
+    uint64_t    bitlen;
+    uint32_t    state[8];
 } sha256_ctx_t;
 
 /*********************** FUNCTION DECLARATIONS **********************/
@@ -60,7 +60,7 @@ int sha256_init(sha256_ctx_t *ctx);
  * \return       更新分組是否成功，成功返回0
  */
 int sha256_update(  sha256_ctx_t    *ctx,
-                    const BYTE      data[],
+                    const uint8_t   data[],
                     size_t          len);
 
 /**
@@ -73,7 +73,7 @@ int sha256_update(  sha256_ctx_t    *ctx,
  * \return       算法輸出結果是否成功，成功返回0
  */
 int sha256_final(   sha256_ctx_t    *ctx,
-                    BYTE            hash[]);
+                    uint8_t         hash[]);
 #ifdef __cplusplus
 }
 #endif
